Refactor most Trilinos wrappers to use Tpetra and related packages
All use of Epetra classes have been replaced with their Tpetra analogs. Solver
and preconditioner packages have been replaced with analogs that play nicer with
Tpetra classes. The non-linear solvers (NOX/LOCA/JFNK) stuff is disabled, since
we dont really use it right now. Will revisit once the linear and eigenvalue
stuff works
