+++ cmake/Dependencies.cmake
@@ -8,7 +8,7 @@
 #++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++!
 
 SET(LIB_REQUIRED_DEP_PACKAGES)
-SET(LIB_OPTIONAL_DEP_PACKAGES Tpetra Epetra Ifpack ML Anasazi Belos NOX CTeuchos ForTeuchos)
+SET(LIB_OPTIONAL_DEP_PACKAGES Tpetra MueLu Ifpack2 Epetra Ifpack ML Anasazi Belos NOX CTeuchos ForTeuchos)
 SET(TEST_REQUIRED_DEP_PACKAGES)
 SET(TEST_OPTIONAL_DEP_PACKAGES)
 SET(LIB_REQUIRED_DEP_TPLS)
