+++ src/trilinos_interfaces/trilinos_c_interfaces.cpp
@@ -10,9 +10,8 @@
 #include "mpi.h"
 #endif
 #ifdef FUTILITY_HAVE_Trilinos
-#include "trilinos_anderson.hpp"
 #include "trilinos_mat_vec.hpp"
-#include "trilinos_tpetra.hpp"
+//#include "trilinos_anderson.hpp"
 #include "trilinos_pc.hpp"
 #include "trilinos_solvers.hpp"
 #include <omp.h>
@@ -20,13 +19,13 @@
 #include "CTeuchos_ParameterList_Cpp.hpp"
 
 bool futility_trilinos_isinit = false;
-Teuchos::RCP< EpetraVecStore > evec(new EpetraVecStore);
-Teuchos::RCP< EpetraMatStore > emat(new EpetraMatStore);
+Teuchos::RCP< TpetraVecStore > tvec(new TpetraVecStore);
+Teuchos::RCP< TpetraMatStore > tmat(new TpetraMatStore);
+Teuchos::RCP< PCStore        > pcst(new PCStore);
 Teuchos::RCP< AnasaziStore   > aeig(new AnasaziStore);
 Teuchos::RCP< BelosStore     > bels(new BelosStore);
-Teuchos::RCP< AndersonStore  > andr(new AndersonStore);
-Teuchos::RCP< JFNKStore      > jfnk(new JFNKStore);
-Teuchos::RCP< PCStore        > pcst(new PCStore);
+// Teuchos::RCP< AndersonStore  > andr(new AndersonStore);
+// Teuchos::RCP< JFNKStore      > jfnk(new JFNKStore);
 
 //------------------------------------------------------------------------------
 // Vector
@@ -34,80 +33,78 @@ Teuchos::RCP< PCStore        > pcst(new PCStore);
 extern "C" void ForPETRA_VecInit(int &id, const int n, const int nlocal,
                                  const int Comm) {
 #ifdef HAVE_MPI
-    id = evec->new_data(n, nlocal, MPI_Comm_f2c(Comm));
+    id = tvec->new_data(n, nlocal, MPI_Comm_f2c(Comm));
 #else
-    id = evec->new_data(n, nlocal, Comm);
+    id = tvec->new_data(n, nlocal, Comm);
 #endif
 }
 
 extern "C" void ForPETRA_VecSetImportMap(const int id, const int n,
                                          const int *gids) {
-    int ierr = evec->define_map_data(id, n, gids);
+    int ierr = tvec->define_map_data(id, n, gids);
     assert(ierr == 0);
 }
 
 extern "C" void ForPETRA_VecDestroy(const int id) {
-    evec->delete_data(id);
+    tvec->delete_data(id);
 }
 
 extern "C" void ForPETRA_VecSet(const int id, const int i, const double val) {
-    int ierr = evec->set_data(id, &i, &val);
-    assert(ierr == 0);
+    tvec->set_data(id, &i, &val);
 }
 
 extern "C" void ForPETRA_VecSetAll(const int id, const double val) {
-    int ierr = evec->setall_data(id, val);
-    assert(ierr == 0);
+    tvec->setall_data(id, val);
 }
 
 extern "C" void ForPETRA_VecTransfer(const int id) {
-    evec->transfer_data(id);
+    tvec->transfer_data(id);
 }
 
 extern "C" void ForPETRA_VecGet(const int id, const int i, double &val) {
-    int ierr = evec->get_data(id, i, val);
+    int ierr = tvec->get_data(id, i, val);
     assert(ierr == 0);
 }
 
 extern "C" void ForPETRA_VecCopy(const int id, const int idfrom) {
-    int ierr = evec->copy_data(id, idfrom);
+    int ierr = tvec->copy_data(id, idfrom);
     assert(ierr == 0);
 }
 
 extern "C" void ForPETRA_VecAXPY(const int id, const int idx, const double a,
                                  const double b) {
-    int ierr = evec->axpy_data(id, idx, a, b);
+    int ierr = tvec->axpy_data(id, idx, a, b);
     assert(ierr == 0);
 }
 
 extern "C" void ForPETRA_VecSum(const int id, double &val) {
     double vals[1];
-    int ierr = evec->norm1_data(id, vals);
+    int ierr = tvec->norm1_data(id, vals);
     val      = vals[0];
     assert(ierr == 0);
 }
 
 extern "C" void ForPETRA_VecNorm2(const int id, double &val) {
     double vals[1];
-    int ierr = evec->norm2_data(id, vals);
+    int ierr = tvec->norm2_data(id, vals);
     val      = vals[0];
     assert(ierr == 0);
 }
 
 extern "C" void ForPETRA_VecMax(const int id, double &val) {
     double vals[1];
-    int ierr = evec->max_data(id, vals);
+    int ierr = tvec->max_data(id, vals);
     val      = vals[0];
     assert(ierr == 0);
 }
 
 extern "C" void ForPETRA_VecScale(const int id, const double val) {
-    int ierr = evec->scale_data(id, val);
+    int ierr = tvec->scale_data(id, val);
     assert(ierr == 0);
 }
 
 extern "C" void ForPETRA_VecEdit(const int id, const char name[]) {
-    int ierr = evec->edit_data(id, name);
+    int ierr = tvec->edit_data(id, name);
     assert(ierr == 0);
 }
 
@@ -117,51 +114,51 @@ extern "C" void ForPETRA_VecEdit(const int id, const char name[]) {
 extern "C" void ForPETRA_MatInit(int &id, const int n, const int nlocal,
                                  const int rnnz, const int Comm) {
 #ifdef HAVE_MPI
-    id = emat->new_data(n, nlocal, rnnz, MPI_Comm_f2c(Comm));
+    id = tmat->new_data(n, nlocal, rnnz, MPI_Comm_f2c(Comm));
 #else
-    id = emat->new_data(n, nlocal, rnnz, Comm);
+    id = tmat->new_data(n, nlocal, rnnz, Comm);
 #endif
 }
 
 extern "C" void ForPETRA_MatDestroy(const int id) {
-    emat->delete_data(id);
+    tmat->delete_data(id);
 }
 
 extern "C" void ForPETRA_MatReset(const int id) {
-    emat->reset_data(id);
+    tmat->reset_data(id);
 }
 
 extern "C" void ForPETRA_MatSet(const int id, const int i, const int nnz,
                                 const int j[], const double val[]) {
-    int ierr = emat->set_data(id, i, nnz, j, val);
+    int ierr = tmat->set_data(id, i, nnz, j, val);
     assert(ierr == 0);
 }
 
 extern "C" void ForPETRA_MatAssemble(const int id) {
-    int ierr = emat->assemble_data(id);
+    int ierr = tmat->assemble_data(id);
     assert(ierr == 0);
 }
 
 extern "C" void ForPETRA_MatGet(const int id, const int i, const int j,
                                 double &val) {
-    int ierr = emat->get_data(id, i, j, val);
+    int ierr = tmat->get_data(id, i, j, val);
     assert(ierr == 0);
 }
 
 extern "C" void ForPETRA_MatMult(const int idA, const bool trans, const int idX,
                                  const int idY) {
     int ierr =
-        emat->matvec_data(idA, trans, evec->get_vec(idX), evec->get_vec(idY));
+        tmat->matvec_data(idA, trans, tvec->get_vec(idX), tvec->get_vec(idY));
     assert(ierr == 0);
 }
 
 extern "C" void ForPETRA_MatEdit(const int id, const char name[]) {
-    int ierr = emat->edit_data(id, name);
+    int ierr = tmat->edit_data(id, name);
     assert(ierr == 0);
 }
 
 extern "C" void ForPETRA_MatNormF(const int id, double &val) {
-    int ierr = emat->normF_data(id, val);
+    int ierr = tmat->normF_data(id, val);
     assert(ierr == 0);
 }
 
@@ -185,7 +182,7 @@ extern "C" void Anasazi_Destroy(const int id) {
 }
 
 extern "C" void Anasazi_SetMat(const int id, const int idLHS, const int idRHS) {
-    aeig->setMat_data(id, emat->get_mat(idLHS), emat->get_mat(idRHS));
+    aeig->setMat_data(id, tmat->get_mat(idLHS), tmat->get_mat(idRHS));
 }
 
 extern "C" void Anasazi_SetPC(const int id, const int idpc) {
@@ -193,7 +190,7 @@ extern "C" void Anasazi_SetPC(const int id, const int idpc) {
 }
 
 extern "C" void Anasazi_SetX(const int id, const int idX) {
-    aeig->setX0_data(id, evec->get_vec(idX));
+    aeig->setX0_data(id, tvec->get_vec(idX));
 }
 
 extern "C" void Anasazi_SetConvCrit(const int id, const double tol,
@@ -247,7 +244,7 @@ extern "C" void Belos_Destroy(const int id) {
 }
 
 extern "C" void Belos_SetMat(const int id, const int idA) {
-    bels->setMat_data(id, emat->get_mat(idA));
+    bels->setMat_data(id, tmat->get_mat(idA));
 }
 
 extern "C" void Belos_SetPC(const int id, const int idpc) {
@@ -255,11 +252,11 @@ extern "C" void Belos_SetPC(const int id, const int idpc) {
 }
 
 extern "C" void Belos_SetX(const int id, const int idX) {
-    bels->setX0_data(id, evec->get_vec(idX));
+    bels->setX0_data(id, tvec->get_vec(idX));
 }
 
 extern "C" void Belos_Setb(const int id, const int idb) {
-    bels->setb_data(id, evec->get_vec(idb));
+    bels->setb_data(id, tvec->get_vec(idb));
 }
 
 extern "C" void Belos_SetConvCrit(const int id, const double tol,
@@ -302,12 +299,12 @@ extern "C" void Preconditioner_Destroy(const int id) {
 }
 
 extern "C" void Preconditioner_Setup(const int id, const int idM) {
-    pcst->setupPC_data(id, emat->get_mat(idM));
+    pcst->setupPC_data(id, tmat->get_mat(idM));
     // std::cout << (pcst->get_pc(pid))->Label() << std::endl;
 }
 
 extern "C" void Preconditioner_Reset(const int id, const int idM) {
-    pcst->resetPC_data(id, emat->get_mat(idM));
+    pcst->resetPC_data(id, tmat->get_mat(idM));
     // std::cout << (pcst->get_pc(pid))->Label() << std::endl;
 }
 
@@ -316,34 +313,40 @@ extern "C" void Preconditioner_Reset(const int id, const int idM) {
 //------------------------------------------------------------------------------
 extern "C" void Anderson_Init(int &id, const int depth, const double beta,
                               const int start, const int idv) {
-    id = andr->new_data(depth, beta, start, evec->get_vec(idv));
+    assert(false);
+    // id = andr->new_data(depth, beta, start, tvec->get_vec(idv));
 }
 
 extern "C" void Anderson_Destroy(const int id) {
-    andr->delete_data(id);
+    assert(false);
+    // andr->delete_data(id);
 }
 
 extern "C" void Anderson_Update(const int id) {
-    andr->step(id);
+    assert(false);
+    // andr->step(id);
 }
 
 extern "C" void Anderson_Reset(const int id) {
-    andr->reset_data(id);
+    assert(false);
+    // andr->reset_data(id);
 }
-
 //------------------------------------------------------------------------------
 // JFNK NOX
 //------------------------------------------------------------------------------
 extern "C" void JFNK_Init(int &id, void (*funptr)(), const int idx,
                           const int idF) {
-    id = jfnk->new_data(funptr, evec->get_vec(idx), evec->get_vec(idF));
+    assert(false);
+    // id = jfnk->new_data(funptr, tvec->get_vec(idx), tvec->get_vec(idF));
 }
 
 extern "C" void JFNK_Destroy(const int id) {
-    jfnk->delete_data(id);
+    assert(false);
+    // jfnk->delete_data(id);
 }
 
 extern "C" void Anderson_Solve(const int id) {
-    jfnk->solve(id);
+    assert(false);
+    // jfnk->solve(id);
 }
 #endif
